Gas optimization, impersonate owner, UIInfo
Gas optimization: I'm using a lower gas default in truffle.config
and setting the needed gas for those deploy transactions that need it
Impersonate owner: In oracle manager the owner of an oracle can be
impersonated by gobernanza (to unregister, remove, etc).
UIInfoGetter: A contract that gives all the information that the UI
need in one call (uggly fix to the fact that rsk node dosen't have
http keep-alive).